Add support for longer strips in SetLEDAdditive

+++ Core/Inc/WS2812.h
@@ -49,7 +49,7 @@ typedef struct comet
 
 // Base LED functions
 void WS2812_SetLED(uint16_t index, uint8_t red, uint8_t green, uint8_t blue);
-void WS2812_SetLEDAdditive(uint8_t index, uint8_t red, uint8_t green, uint8_t blue);
+void WS2812_SetLEDAdditive(uint16_t index, uint8_t red, uint8_t green, uint8_t blue);
 void WS2812_SetAllLEDs(uint32_t red, uint32_t green, uint32_t blue);
 void WS2812_ClearLEDs(void);
 void WS2812_FadeAll(uint8_t denominator);
