+++ Core/Src/WS2812.c
@@ -41,9 +41,9 @@ void WS2812_SetLED(uint16_t index, uint8_t red, uint8_t green, uint8_t blue)
 }
 
 // Adds the RGB color values to the LED at index in the LEDData buffer
-void WS2812_SetLEDAdditive(uint8_t index, uint8_t red, uint8_t green, uint8_t blue)
+void WS2812_SetLEDAdditive(uint16_t index, uint8_t red, uint8_t green, uint8_t blue)
 {
-	LEDData[index][0] = (LEDData[index][0]  + red) > UINT8_MAX ? 255 : (LEDData[index][0] + red);
+	LEDData[index][0] = (LEDData[index][0] + red) > UINT8_MAX ? 255 : (LEDData[index][0] + red);
 	LEDData[index][1] = (LEDData[index][1] + green) > UINT8_MAX ? 255 : (LEDData[index][1] + green);
 	LEDData[index][2] = (LEDData[index][2] + blue) > UINT8_MAX ? 255 : (LEDData[index][2] + blue);
 }
